2008-06-05  Tatsuhiro Tsujikawa  <tujikawa at rednoah dot com>
	Made `begin' unsigned int.
	* src/GrowSegment.h
	* src/Piece.cc
	* src/Piece.h
	* src/PiecedSegment.cc
	* src/PiecedSegment.h
	* src/Segment.h

+++ src/GrowSegment.h
@@ -92,7 +92,7 @@ public:
 
 #ifdef ENABLE_MESSAGE_DIGEST
 
-  virtual bool updateHash(size_t begin,
+  virtual bool updateHash(uint32_t begin,
 			  const unsigned char* data, size_t dataLength)
   {
     return false;
