+++ src/PiecedSegment.h
@@ -85,7 +85,7 @@ public:
 #ifdef ENABLE_MESSAGE_DIGEST
 
   // `begin' is a offset inside this segment.
-  virtual bool updateHash(size_t begin,
+  virtual bool updateHash(uint32_t begin,
 			  const unsigned char* data, size_t dataLength);
 
   virtual bool isHashCalculated() const;
